remove indirection to submodule `dashing.dashing`
